Refactor MegaStore Contents
